Renamed methods in UrlBuilder. Used UrlBuilder as fluent interface.
It turned out that UrlBuilder was offering a fluent interface already,
but the methods indicated that it should be used in a standard way. I
therefore changed the method names & adjusted the callers to use it as
a fluent interface. I also create a uriParameter method that
encodes the URI parameter value before adding it.

Closes #29